requirements.txt
anthropic>=0.42.0
python-telegram-bot[job-scheduler]>=21.0
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
google-api-python-client>=2.160.0
google-auth-oauthlib>=1.2.0
